requirements.txt
opencv-python
pyserial
Pillow
numpy
scikit-image
rpi_ws281x